+++ packages/ts/jns42-generator/src/generators/main-ts.ts
@@ -1,8 +1,13 @@
+import { nestedTextFromTs } from "../utils/index.js";
 import { CodeGeneratorBase } from "./code-generator-base.js";
 import { TypesTsCodeGenerator } from "./types-ts.js";
 import { ValidatorsTsCodeGenerator } from "./validators-ts.js";
 
 export class MainTsCodeGenerator extends CodeGeneratorBase {
+  public *getCode() {
+    yield nestedTextFromTs(this.factory, this.getStatements());
+  }
+
   public *getStatements() {
     const { factory: f } = this;
