+++ packages/ts/jns42-generator/src/utils/format.ts
@@ -1,10 +1,10 @@
-import ts from "typescript";
+import { NestedText, flattenNestedText } from "./iterable-text-template.js";
 
 export function formatData(content: unknown) {
   return JSON.stringify(content, undefined, 2);
 }
 
-export function formatStatements(factory: ts.NodeFactory, statements: Iterable<ts.Statement>) {
+export function formatStatements(nestedCode: NestedText) {
   const banner = `
 // @generated by
 //     __             _____     _                 ___ ___ 
@@ -14,15 +14,8 @@ export function formatStatements(factory: ts.NodeFactory, statements: Iterable<t
 //                                 -- www.JsonSchema42.org
 `.trim();
 
-  const printer = ts.createPrinter({
-    newLine: ts.NewLineKind.LineFeed,
-  });
+  const code = [...flattenNestedText(nestedCode)].join("");
+  const formattedCode = code.toString();
 
-  const sourceFile = factory.createSourceFile(
-    [...statements],
-    factory.createToken(ts.SyntaxKind.EndOfFileToken),
-    ts.NodeFlags.None,
-  );
-
-  return `${banner}\n\n${printer.printFile(sourceFile)}`;
+  return `${banner}\n\n${formattedCode}`;
 }
