+++ packages/ts/jns42-generator/src/utils/index.ts
@@ -1,5 +1,5 @@
 export * from "./choose.js";
-export * from "./file.js";
+export * from "./format.js";
 export * from "./iterable-text-template.js";
 export * from "./literal.js";
 export * from "./load.js";
