+++ packages/ts/jns42-generator/src/utils/iterable-text-template.ts
@@ -1,3 +1,5 @@
+import ts from "typescript";
+
 export type NestedText = Iterable<NestedText> | string;
 
 export function* iterableTextTemplate(
@@ -24,4 +26,21 @@ export function* flattenNestedText(nestedText: NestedText): Iterable<string> {
       yield* flattenNestedText(text);
     }
   }
+}
+
+/**
+ * temporary helper function, should be factored out
+ */
+export function nestedTextFromTs(factory: ts.NodeFactory, statements: Iterable<ts.Statement>) {
+  const printer = ts.createPrinter({
+    newLine: ts.NewLineKind.LineFeed,
+  });
+
+  const sourceFile = factory.createSourceFile(
+    [...statements],
+    factory.createToken(ts.SyntaxKind.EndOfFileToken),
+    ts.NodeFlags.None,
+  );
+
+  return printer.printFile(sourceFile);
 }
